liberwiki/requirements.txt
wheel==0.44.0
Django==5.1
iubeo==0.3.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3
sentry-sdk==2.13.0
django-lifecycle==1.2.4
django-pghistory==3.2.0
django-extensions==3.2.3
ipython==8.26.0
django-admin-interface==0.28.8
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
django-cors-headers==4.4.0
memoization==0.4.0
django-hosts==6.0
pyuwsgi==2.0.26
django-pgtrigger==4.12.2
whitenoise==6.7.0
mjml-python==1.3.4
django-allauth[socialaccount]==65.6.0
django-more-admin-filters==1.12
